requirements.txt
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==2.1.1
custom-inherit==2.4.0
decorator==5.1.1
enum34==1.1.10
idna==3.4
markovchain==0.2.6
Mastodon.py==1.8.0
python-dateutil==2.8.2
python-magic==0.4.27
PyYAML==6.0
requests==2.28.1
six==1.16.0
tqdm==4.64.1
urllib3==1.26.13
